# Torch Generator Agent