… docs/source/torch_agent.rst
… parlai/agents/legacy_agents/seq2seq/dict_v1.py
… parlai/agents/legacy_agents/seq2seq/modules_v1.py
… parlai/agents/legacy_agents/seq2seq/seq2seq_v1.py
… parlai/agents/legacy_agents/seq2seq/torch_agent_v1.py
… parlai/agents/seq2seq/seq2seq.py
… parlai/core/torch_agent.py
… parlai/core/torch_generator_agent.py
… parlai/core/torch_ranker_agent.py
… parlai/core/utils.py
… tests/test_seq2seq.py