Checkpoint, got beam search generalized so it works with transformer (another PR). Has the tradeoffs we discussed.

Going to have so many merge problems...