@alexholdenmiller 2018-10-31
Overall looks pretty good, although kind of sad that it looks like decoding is now O(n^2) instead of O(n) for seq2seq :(

Something like this could make it O(n) again and would enable attention over currently produced tokens, although now it's a different model than it was before (but could be better):
```
 def decode(self, bsz, encoder_states, maxlen):
        """Greedy search"""
        xs = self._starts(bsz)
        for _ in range(maxlen):
            output = self.decoder(xs, encoder_states)
            scores = output[:, -1, :]
            _, preds = scores.max(dim=-1)
            xs = preds
            encoder_states = ( torch.cat((encoder_states[0], output)), output, torch.cat(attn_mask, torch.ones())
        return xs
```