@stephenroller 2018-10-31
Yeah it's problematic. I hoped that it wouldn't be too bad, but I found that some actual testing of it found that beam search was untenable this way.

I'm going to implement incremental state today, so that the hidden state is cached.