@stephenroller 2018-11-02
Alright, incremental decoding is back, so all is well.

Transformer won't support incremental decoding initially.