@stephenroller 2018-11-02
Actually, right now it's surprisingly a *lot* faster, but that may be to #1262 .