@stephenroller 2018-11-02
Speedup in this example probably comes from skipping ranking candidates when it's not explicitly asked for.

Before:
```
[ training... ]
preinitializing pytorch cuda buffer
[ time:2s total_exs:3808 epochs:0.42 time_left:22s ] {'exs': 3808, 'token_acc': 0.527, 'loss': 1.573, 'ppl': 4.823, 'num_epochs': 0}
[ time:4s total_exs:8320 epochs:0.92 time_left:17s ] {'exs': 4512, 'token_acc': 0.5836, 'loss': 0.9189, 'ppl': 2.506, 'num_epochs': 0}
[ running eval: valid ]
[creating task(s): babi:task10k:1]
[loading fbdialog data:/private/home/roller/working/parlai/data/bAbI/tasks_1-20_v1-2/en-valid-10k-nosf/qa1_valid.txt]
valid:{'exs': 1000, 'accuracy': 0.169, 'f1': 0.169, 'bleu': 1.69e-10, 'token_acc': 0.5845, 'loss': 0.9164, 'ppl': 2.5}
[ new best accuracy: 0.169 ]
[ time:7s total_exs:9056 epochs:1.01 time_left:30s ] {'exs': 736, 'token_acc': 0.6094, 'loss': 0.8758, 'ppl': 2.401, 'num_epochs': 1}
[ time:9s total_exs:13312 epochs:1.48 time_left:22s ] {'exs': 4256, 'token_acc': 0.5874, 'loss': 0.9082, 'ppl': 2.48, 'num_epochs': 1}
[ time:11s total_exs:17440 epochs:1.94 time_left:18s ] {'exs': 4128, 'token_acc': 0.5907, 'loss': 0.9052, 'ppl': 2.472, 'num_epochs': 1}
[ running eval: valid ]
valid:{'exs': 1000, 'accuracy': 0.169, 'f1': 0.169, 'bleu': 1.69e-10, 'token_acc': 0.5845, 'loss': 0.9083, 'ppl': 2.48}
[ did not beat best accuracy: 0.169 impatience: 1 ]
[ time:15s total_exs:18080 epochs:2.01 time_left:22s ] {'exs': 640, 'token_acc': 0.5781, 'loss': 0.8969, 'ppl': 2.452, 'num_epochs': 2}
[ time:17s total_exs:22400 epochs:2.49 time_left:17s ] {'exs': 4320, 'token_acc': 0.585, 'loss': 0.9055, 'ppl': 2.473, 'num_epochs': 2}
[ time:19s total_exs:26688 epochs:2.97 time_left:13s ] {'exs': 4288, 'token_acc': 0.5842, 'loss': 0.9039, 'ppl': 2.469, 'num_epochs': 2}
[ running eval: valid ]
valid:{'exs': 1000, 'accuracy': 0.155, 'f1': 0.155, 'bleu': 1.55e-10, 'token_acc': 0.5775, 'loss': 0.9065, 'ppl': 2.476}
[ did not beat best accuracy: 0.169 impatience: 2 ]
[ time:22s total_exs:27104 epochs:3.01 time_left:14s ] {'exs': 416, 'token_acc': 0.5625, 'loss': 0.8941, 'ppl': 2.445, 'num_epochs': 3}
[ time:24s total_exs:31424 epochs:3.49 time_left:10s ] {'exs': 4320, 'token_acc': 0.5856, 'loss': 0.9038, 'ppl': 2.469, 'num_epochs': 3}
[ time:26s total_exs:35776 epochs:3.98 time_left:6s ] {'exs': 4352, 'token_acc': 0.5944, 'loss': 0.8988, 'ppl': 2.457, 'num_epochs': 3}
[ running eval: valid ]
valid:{'exs': 1000, 'accuracy': 0.259, 'f1': 0.259, 'bleu': 2.59e-10, 'token_acc': 0.6295, 'loss': 0.8676, 'ppl': 2.381}
[ new best accuracy: 0.259 (previous best was 0.169) ]
[ time:29s total_exs:36128 epochs:4.01 time_left:7s ] {'exs': 352, 'token_acc': 0.5938, 'loss': 0.8676, 'ppl': 2.381, 'num_epochs': 4}
[ time:31s total_exs:40448 epochs:4.49 time_left:3s ] {'exs': 4320, 'token_acc': 0.6197, 'loss': 0.8653, 'ppl': 2.376, 'num_epochs': 4}
[ time:33s total_exs:44736 epochs:4.97 time_left:0s ] {'exs': 4288, 'token_acc': 0.6245, 'loss': 0.8483, 'ppl': 2.336, 'num_epochs': 4}
[ time:33s total_exs:45024 epochs:5.0 time_left:-1s ] {'exs': 288, 'token_acc': 0.6337, 'loss': 0.8446, 'ppl': 2.327, 'num_epochs': 5}
[ num_epochs completed:5.0 time elapsed:33.85872149467468s ]
[ running eval: valid ]
[creating task(s): babi:task10k:1]
[loading fbdialog data:/private/home/roller/working/parlai/data/bAbI/tasks_1-20_v1-2/en-valid-10k-nosf/qa1_valid.txt]
valid:{'exs': 1000, 'accuracy': 0.312, 'f1': 0.312, 'bleu': 3.12e-10, 'token_acc': 0.656, 'loss': 0.8158, 'ppl': 2.261}
[ running eval: test ]
[creating task(s): babi:task10k:1]
[loading fbdialog data:/private/home/roller/working/parlai/data/bAbI/tasks_1-20_v1-2/en-valid-10k-nosf/qa1_test.txt]
test:{'exs': 1000, 'accuracy': 0.272, 'f1': 0.272, 'bleu': 2.72e-10, 'token_acc': 0.636, 'loss': 0.8446, 'ppl': 2.327}
python -u examples/train_model.py -t babi:task10k:1 --dict-file  -bs 32 -dt    42.34s user 4.66s system 106% cpu 44.201 total
```

After:
```
[ training... ]
[ time:2s total_exs:3296 epochs:0.37 time_left:25s ] {'exs': 3296, 'token_acc': 0.5155, 'loss': 1.695, 'ppl': 5.448, 'num_epochs': 0}
[ time:4s total_exs:7008 epochs:0.78 time_left:22s ] {'exs': 3712, 'token_acc': 0.5789, 'loss': 0.9241, 'ppl': 2.52, 'num_epochs': 0}
[ running eval: valid ]
[creating task(s): babi:task10k:1]
[loading fbdialog data:/private/home/roller/working/parlai/data/bAbI/tasks_1-20_v1-2/en-valid-10k-nosf/qa1_valid.txt]
valid:{'exs': 1000, 'accuracy': 0.169, 'f1': 0.169, 'bleu': 1.69e-10, 'token_acc': 0.5845, 'loss': 0.916, 'ppl': 2.499}
[ new best accuracy: 0.169 ]
[ time:6s total_exs:9696 epochs:1.08 time_left:22s ] {'exs': 2688, 'token_acc': 0.5878, 'loss': 0.9081, 'ppl': 2.48, 'num_epochs': 1}
[ time:8s total_exs:13888 epochs:1.54 time_left:18s ] {'exs': 4192, 'token_acc': 0.5839, 'loss': 0.9085, 'ppl': 2.481, 'num_epochs': 1}
[ running eval: valid ]
valid:{'exs': 1000, 'accuracy': 0.169, 'f1': 0.169, 'bleu': 1.69e-10, 'token_acc': 0.5845, 'loss': 0.9061, 'ppl': 2.475}
[ did not beat best accuracy: 0.169 impatience: 1 ]
[ time:10s total_exs:18080 epochs:2.01 time_left:15s ] {'exs': 4192, 'token_acc': 0.5938, 'loss': 0.8901, 'ppl': 2.435, 'num_epochs': 2}
[ time:12s total_exs:22240 epochs:2.47 time_left:12s ] {'exs': 4160, 'token_acc': 0.5894, 'loss': 0.9026, 'ppl': 2.466, 'num_epochs': 2}
[ time:14s total_exs:26048 epochs:2.89 time_left:10s ] {'exs': 3808, 'token_acc': 0.5961, 'loss': 0.8984, 'ppl': 2.456, 'num_epochs': 2}
[ running eval: valid ]
valid:{'exs': 1000, 'accuracy': 0.173, 'f1': 0.173, 'bleu': 1.73e-10, 'token_acc': 0.5865, 'loss': 0.8808, 'ppl': 2.413}
[ new best accuracy: 0.173 (previous best was 0.169) ]
[ time:16s total_exs:28736 epochs:3.19 time_left:9s ] {'exs': 2688, 'token_acc': 0.6205, 'loss': 0.8692, 'ppl': 2.385, 'num_epochs': 3}
[ time:18s total_exs:32736 epochs:3.64 time_left:6s ] {'exs': 4000, 'token_acc': 0.623, 'loss': 0.8549, 'ppl': 2.351, 'num_epochs': 3}
[ running eval: valid ]
valid:{'exs': 1000, 'accuracy': 0.27, 'f1': 0.27, 'bleu': 2.7e-10, 'token_acc': 0.635, 'loss': 0.8156, 'ppl': 2.26}
[ new best accuracy: 0.27 (previous best was 0.173) ]
[ time:20s total_exs:36128 epochs:4.01 time_left:5s ] {'exs': 3392, 'token_acc': 0.6562, 'loss': 0.8144, 'ppl': 2.258, 'num_epochs': 4}
[ time:22s total_exs:40224 epochs:4.47 time_left:2s ] {'exs': 4096, 'token_acc': 0.6355, 'loss': 0.833, 'ppl': 2.3, 'num_epochs': 4}
[ time:24s total_exs:44032 epochs:4.89 time_left:0s ] {'exs': 3808, 'token_acc': 0.6364, 'loss': 0.8255, 'ppl': 2.283, 'num_epochs': 4}
[ time:25s total_exs:45024 epochs:5.0 time_left:-1s ] {'exs': 992, 'token_acc': 0.6628, 'loss': 0.8111, 'ppl': 2.25, 'num_epochs': 5}
[ num_epochs completed:5.0 time elapsed:25.180065393447876s ]
[ running eval: valid ]
[creating task(s): babi:task10k:1]
[loading fbdialog data:/private/home/roller/working/parlai/data/bAbI/tasks_1-20_v1-2/en-valid-10k-nosf/qa1_valid.txt]
valid:{'exs': 1000, 'accuracy': 0.378, 'f1': 0.378, 'bleu': 3.78e-10, 'token_acc': 0.689, 'loss': 0.7786, 'ppl': 2.178}
[ running eval: test ]
[creating task(s): babi:task10k:1]
[loading fbdialog data:/private/home/roller/working/parlai/data/bAbI/tasks_1-20_v1-2/en-valid-10k-nosf/qa1_test.txt]
test:{'exs': 1000, 'accuracy': 0.329, 'f1': 0.329, 'bleu': 3.29e-10, 'token_acc': 0.6645, 'loss': 0.8166, 'ppl': 2.263}
python examples/train_model.py -t babi:task10k:1 --dict-file  -bs 32 -dt  -ep  29.04s user 4.87s system 109% cpu 30.981 total
tee got.log  0.00s user 0.00s system 0% cpu 30.980 total
```

The accuracies are a little different, but pretty sure that's just random. The changes are actually binary compatible, so I've been taking a model trained on master branch and ensuring it produces the same results as this model, and vice versa.